manipulate middleware with app.use
