+++ 02-express-tutorial/app.js
@@ -1,20 +1,15 @@
 const express = require('express')
+const logger = require('./logger')
 
 const app = express()
+// req => middleware => res
+app.use(logger)
 
-const logger = (req, res, next) => {
-  const method = req.method
-  const url = req.url
-  const time = new Date().getFullYear()
-  console.log(method, url, time)
-  next()
-}
-
-app.get('/', logger, (req, res) => {
+app.get('/', (req, res) => {
   res.send('Home page')
 })
 
-app.get('/about', logger, (req, res) => {
+app.get('/about', (req, res) => {
   res.send('About page')
 })
